pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snaptex"
version = "1.0.0"
description = "Screenshot to LaTeX/Text recognizer for Windows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.12"
authors = [
    { name = "SnapTeX Contributors" },
]
keywords = ["ocr", "latex", "formula", "screenshot", "windows"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pix2text>=1.1.6",
    "PySide6>=6.5",
    "PySide6-Addons>=6.5",
    "rapidocr-onnxruntime>=1.4",
    "Pillow",
    "numpy",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0",
]
dev = [
    "pyinstaller>=6.0",
    "pytest",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/SnapTeX"
Issues = "https://github.com/YOUR_USERNAME/SnapTeX/issues"

[project.gui-scripts]
snaptex = "snaptex.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/snaptex"]
